nit: extract helper functions into separate file
